Fixes for compilation on x86_64.
Note, if you want to cast a pointer to an integer it is best to use

#include <stdint.h>
int x = (int)(intptr_t)p;  // where p is your pointer

to avoid all the warnings.

If you are printing out a pointer %p works.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2315 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/analysis/fa.cpp
@@ -415,7 +415,7 @@ type_cannonicalize(AType *t) {
     qsort_pointers((void**)&t->sorted.v[0], (void**)t->sorted.end());
   unsigned int h = 0;
   for (int i = 0; i < t->sorted.n; i++)
-    h = (uint)t->sorted.v[i] * open_hash_multipliers[i % 256];
+    h = (uint)(intptr_t)t->sorted.v[i] * open_hash_multipliers[i % 256];
   t->hash = h ? h : h + 1; // 0 is empty
   AType *tt = cannonical_atypes.put(t);
   if (!tt) tt = t;
@@ -2581,7 +2581,7 @@ setters_cannonicalize(Setters *s) {
     qsort_pointers((void**)&s->sorted.v[0], (void**)s->sorted.end());
   uint h = 0;
   for (int i = 0; i < s->sorted.n; i++)
-    h = (uint)s->sorted.v[i] * open_hash_multipliers[i % 256];
+    h = (uint)(intptr_t)s->sorted.v[i] * open_hash_multipliers[i % 256];
   s->hash = h ? h : h + 1; // 0 is empty
   Setters *ss = cannonical_setters.put(s);
   if (!ss) ss = s;
